Move row conversion logic to its own file
The QueryDriver file is quite large and this code is logically separate, so move it to its own file.
